[#11] Remake folder struct
