--- src/modules/v1/services/count/current.service.js
@@ -1,9 +1,0 @@
-const { getCount } = require('../../../../utils/storage/count.storage');
-
-/**
- * Get the current count.
- * @returns {number} The current count
- */
-exports.current = async () => {
-    return getCount();
-};
