--- src/modules/v1/services/count/reset.service.js
@@ -1,9 +1,0 @@
-const { resetCount } = require('../../../../utils/storage/count.storage');
-
-/**
- * Reset the count to 0.
- * @returns {number} The reset count (0)
- */
-exports.reset = async () => {
-    return resetCount();
-};
