+++ src/v1/models/authentication.model.js
@@ -1,0 +1,57 @@
+const { DataTypes } = require("sequelize");
+const sequelize = require("../../configs/database.config");
+
+/** 
+ * @description Authentication Model 
+ * @typedef {Object} Authentication
+ * @property {number} authentication_id - Unique identifier for authentication records.
+ * @property {string} identifier_email - Email identifier for authentication (max 1000 chars).
+ * @property {string} password - Hashed password for authentication.
+ * @property {"ADMIN" | "STORE" | "CUSTOMER"} role - Role of the authenticated user.
+ * @property {Date} created_at - Timestamp for when the record was created.
+ * @property {Date} updated_at - Timestamp for when the record was last updated.
+ * @property {boolean} delete_flag - Soft delete flag (true if deleted).
+ */
+const Authentication = sequelize.define("Authentication", {
+    authentication_id: {
+        type: DataTypes.INTEGER,
+        primaryKey: true,
+        autoIncrement: true,
+    },
+    identifier_email: {
+        type: DataTypes.BLOB,
+        allowNull: false,
+        validate: {
+            len: [0, 1000],
+        },
+        get() {
+            const rawValue = this.getDataValue("identifier_email");
+            return rawValue ? rawValue.toString("utf8") : null;
+        },
+    },
+    password: {
+        type: DataTypes.BLOB,
+        allowNull: false,
+        validate: {
+            len: [0, 1000],
+        },
+        get() {
+            const rawValue = this.getDataValue("password");
+            return rawValue ? rawValue.toString("utf8") : null;
+        },
+    },
+    role: {
+        type: DataTypes.ENUM("ADMIN", "STORE", "CUSTOMER"),
+        allowNull: false,
+    },
+    delete_flag: {
+        type: DataTypes.BOOLEAN,
+        defaultValue: false,
+    },
+}, {
+    tableName: "authentications",
+    timestamps: true,
+    underscored: true,
+});
+
+module.exports = Authentication;
