+++ src/v1/services/count/current.service.js
@@ -1,0 +1,9 @@
+const { getCount } = require('../../utils/storage/count.storage');
+
+/**
+ * Get the current count.
+ * @returns {number} The current count
+ */
+exports.current = async () => {
+    return getCount();
+};
