+++ src/v1/utils/logs/.gitignore
@@ -1,0 +1,2 @@
+*
+!.gitignore
